murderered bugs to the point it compiles

+++ include/maths/vec.h
@@ -6,6 +6,7 @@
 #include <array>
 #include <algorithm>
 #include "core/constant.h"
+#include <iostream>
 
 template<class vec_type_t, unsigned int N>
 class vec_t {
@@ -132,8 +133,8 @@ public:
     }
 };
 
-typedef vec_t<double, 2> vec2_t;
-typedef vec_t<double, 3> vec3_t;
-typedef vec_t<double, 4> vec4_t;
+typedef vec_t<float, 2> vec2_t;
+typedef vec_t<float, 3> vec3_t;
+typedef vec_t<float, 4> vec4_t;
 
 #endif
